setseen overwrite
When syncserver gets a setseen_all, it contains an entire skiplist file
from the master.  What, skiplist you say?  Yeah - I'm afraid so.  It's
sent in binary.  That sucks, but here we have it.  Fixing that would be
a protocol change.

But... at least we can just copy the whole thing into place rather than
fiddling about updating individual fields.  We _know_ it's correct
because the master is always right[tm].