Limit the number of message headers added to the cache
This is in response to a site that had emails with 31,000
copies of the same headers added over and over again by a
malfunctioning program.  The messages caused huge cache
records and a massive IO hit just viewing them, generating
a denial of service.

By having a limit, sane emails won't be affected, and
broken emails should still be reasonably findable because
the first 1000 headers will usually contain the important
information anyway.