statuscache: don't fatal() on missing database
Thanks to Cyril Servant <elfejoyeux@gmail.com>, who said:

I wonder why is a "fatal" function called when the statuscache backend
can't be opened. This makes the process terminate (then fork again,
then terminate, then...). This is not logical, because statuscache
is... a cache ! Cyrus should work fine without (and actually, it does,
with the patch attached).

As we use an sql backend for statuscache, when the sql server is not
available, we prefer that cyrus-imapd works without cache than a total
breakdown.