[sieve] test.c: prevent crashes when -t or -e are not specified
If a the test program is invoked on a script that uses the envelope
test without using -e and -t command line options, it will crash due to
trying to dereference a null pointer.  The command line options are:
   -e envelope_from
   -t envelope_to

bc_eval.c assumes that getenvelope() will return a valid array of
values.  This patch makes sure that e_from and e_to have non-null data.